code/scripts/db/requirements.txt
pandas==1.2.3
SQLAlchemy==1.4.6
PyYAML==5.4.1
PyMySQL==1.0.2
